requirements.txt
opencv-python==4.5.3.56
Flask==2.1.2
requests==2.28.0
streamlit
streamlit-webrtc
plotly-express==0.4.1
pandas
pyinstaller
